+++ component/Demo/version.json
@@ -1,0 +1,4 @@
+{
+  "version": "3.0.0",
+  "name": "demo"
+}
