+++ component/NotFound/NotFoundTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Neoan3\Components;
+namespace Neoan3\Component\NotFound;
 
 use PHPUnit\Framework\TestCase;
 
@@ -8,7 +8,7 @@ class NotFoundTest extends TestCase
 {
     public function testInit()
     {
-        $st = new NotFound();
+        $st = new NotFoundController();
         $this->expectOutputRegex('/^<!doctype html>/');
         $st->init();
     }
