--- component/demo/version.json
@@ -1,4 +1,0 @@
-{
-  "version": "0.0.1",
-  "name": "demo"
-}
